Added patch handler for a single activity and added error handling to activity post requests.

+++ functions/src/controllers/activityController.ts
@@ -4,13 +4,44 @@ import {
   validateRequestBody,
   validateRequestParams,
 } from "zod-express-middleware";
-import { activateOrDeactivateParamSchema, activityArraySchema } from "../types";
+import {
+  activateOrDeactivateParamSchema,
+  activityArraySchema,
+  activityOptionalSchema,
+} from "../types";
 import { z } from "zod";
 
 export const activityController = Router();
 const prisma = new PrismaClient();
 
+const getAllActivityNames = async () => {
+  return await prisma.activity
+    .findMany({
+      select: {
+        name: true,
+      },
+    })
+    .then((activities) => activities.map((activity) => activity.name));
+};
+
+const verifyNameIsUnique = async (newName: string) => {
+  const allActivityNames = await getAllActivityNames();
+
+  return allActivityNames.every((activityName) => activityName !== newName);
+};
+
+// Get all active activities
 activityController.get("/", async (_req, res) => {
+  const allActiveActivities = await prisma.activity.findMany({
+    where: {
+      isActive: true,
+    },
+  });
+  return res.status(200).json(allActiveActivities);
+});
+
+// Get all activities
+activityController.get("/all", async (_req, res) => {
   const allActivities = await prisma.activity.findMany();
   return res.status(200).json(allActivities);
 });
@@ -19,13 +50,43 @@ activityController.post(
   "/",
   validateRequestBody(activityArraySchema),
   async (req, res) => {
+    // Check if names are unique.
+    const namesAreUniqueArr = await Promise.all(
+      req.body.map(async (activity) => {
+        const nameIsUnique = await verifyNameIsUnique(activity.name);
+        if (!nameIsUnique) {
+          return activity.name;
+        }
+        return true;
+      })
+    );
+
+    const namesAreUnique = namesAreUniqueArr.every((item) => item === true);
+    if (!namesAreUnique) {
+      const invalidActivities = namesAreUniqueArr.filter(
+        (item) => item !== true
+      );
+      return res.status(400).json({
+        message:
+          "No activities created. One or more entered names are already in use.",
+        invalidActivities,
+      });
+    }
+
+    // Create characters.
+    const activityNoIdArr = req.body.map((activityInfo) => {
+      const { id, ...activityNoId } = activityInfo;
+      return activityNoId;
+    });
     try {
       const createdActivities = await prisma.activity.createMany({
-        data: req.body,
+        data: activityNoIdArr,
       });
+      const activityNames = req.body.map((activity) => activity.name);
       return res.status(200).json({
-        message: "Created activities.",
-        activities: createdActivities,
+        message: "Created characters.",
+        count: createdActivities.count,
+        activities: activityNames,
       });
     } catch (error) {
       return res.status(400).json({
@@ -41,6 +102,35 @@ activityController.delete(
   validateRequestBody(activityArraySchema),
   async (req, res) => {
     const names = req.body.map((activity) => activity.name);
+
+    // Verify that none of the activities have a history of being booked.
+    const activityForEventIds = await prisma.activitiesForEvent
+      .findMany()
+      .then((activityForEvent) =>
+        activityForEvent.map((item) => item.activityId)
+      );
+
+    const activitiesNeverBookedArr = req.body.map((activity) => {
+      if (activityForEventIds.includes(activity.id)) {
+        return activity;
+      }
+      return true;
+    });
+    const activitiesNeverBooked = activitiesNeverBookedArr.every(
+      (item) => item === true
+    );
+    if (!activitiesNeverBooked) {
+      const activitiesWithHistory = activityArraySchema.parse(
+        activitiesNeverBookedArr.filter((item) => item !== true)
+      );
+      return res.status(400).json({
+        message:
+          "Unable to delete activities. One or more activities have an event history.",
+        activitiesWithHistory,
+      });
+    }
+
+    // Delete activities.
     try {
       const deletedActivities = await prisma.activity.deleteMany({
         where: {
@@ -50,9 +140,12 @@ activityController.delete(
         },
       });
 
+      const activityNames = req.body.map((activity) => activity.name);
+
       return res.status(200).json({
-        message: "Deleted activities.",
+        message: "Deleted characters.",
         count: deletedActivities.count,
+        activities: activityNames,
       });
     } catch (error) {
       return res.status(400).json({
@@ -60,6 +153,38 @@ activityController.delete(
         error: error instanceof Error ? error.message : error,
       });
     }
+  }
+);
+
+// Update a single activity.
+activityController.patch(
+  "/",
+  validateRequestBody(activityOptionalSchema),
+  async (req, res) => {
+    const activityId = +req.body.id;
+    const { id, ...dataNoId } = req.body;
+
+    // Verify the new name is unique.
+    if (req.body.name) {
+      const nameIsUnique = await verifyNameIsUnique(req.body.name);
+      if (!nameIsUnique) {
+        return res.status(400).json({
+          message: `Activity names must be unique. '${req.body.name}' is already used.`,
+        });
+      }
+    }
+
+    // Update activity.
+    const updatedActivity = await prisma.activity.update({
+      where: {
+        id: activityId,
+      },
+      data: dataNoId,
+    });
+
+    return res
+      .status(200)
+      .json({ message: "Updated activity.", updatedActivity });
   }
 );
 
@@ -78,18 +203,7 @@ activityController.patch(
     const activateOrDeactivate = req.params.activateOrDeactivate === "activate";
 
     // Validate the spelling of all names.
-    const allActivityNames = await prisma.activity
-      .findMany({
-        where: {
-          name: {
-            in: req.body,
-          },
-        },
-        select: {
-          name: true,
-        },
-      })
-      .then((activities) => activities.map((activity) => activity.name));
+    const allActivityNames = await getAllActivityNames();
 
     const allEntriesValidArr = req.body.map((activityName) => {
       if (!allActivityNames.includes(activityName)) {
